making timing struct use single precision

+++ Assets/Scripts/Structs/Timing.cs
@@ -2,10 +2,10 @@
 {
     public struct Timing
     {
-        public double Start;
-        public double End;
+        public float Start;
+        public float End;
 
-        public Timing(double start = 0, double end = 0)
+        public Timing(float start = 0, float end = 0)
         {
             Start = start;
             End = end;
